+++ rotkehlchen/db/upgrades/v37_v38.py
@@ -111,18 +111,34 @@ def _reset_decoded_events(write_cursor: 'DBCursor') -> None:
     log.debug('Enter _reset_decoded_events')
 
 
+def _remove_duplicate_block_mev_rewards(write_cursor: 'DBCursor') -> None:
+    """If mev reward is exact same as block production reward then it's a duplicate event.
+    In that case it needs to be deleted.
+    """
+    log.debug('Enter _remove_duplicate_block_mev_rewards')
+    write_cursor.execute(
+        'DELETE FROM history_events WHERE identifier IN ('
+        'SELECT B.identifier from history_events A INNER JOIN history_events B '
+        'ON A.event_identifier=B.event_identifier AND A.subtype="block production" '
+        'AND B.subtype="mev reward" AND A.amount=B.amount WHERE A.entry_type=4);',
+    )
+    log.debug('Exit _remove_duplicate_block_mev_rewards')
+
+
 def upgrade_v37_to_v38(db: 'DBHandler', progress_handler: 'DBUpgradeProgressHandler') -> None:
     """Upgrades the DB from v37 to v38. This was in v1.29.0 release.
-
         - Reset decoded events
         - Reduce the data stored per internal transaction
         - Add Polygon POS location and nodes
         - Drop the unused aave events
+        - Remove potential duplicate block mev reward events
     """
     log.debug('Entered userdb v37->v38 upgrade')
-    progress_handler.set_total_steps(5)
+    progress_handler.set_total_steps(6)
     with db.user_write() as write_cursor:
         _reset_decoded_events(write_cursor)
+        progress_handler.new_step()
+        _remove_duplicate_block_mev_rewards(write_cursor)
         progress_handler.new_step()
         _reduce_internal_txs(write_cursor)
         progress_handler.new_step()