+++ docs/changelog.rst
@@ -22,6 +22,7 @@ Changelog
 * :feature:`-` The PnL report page will load faster if there is many old reports in the DB.
 * :feature:`-` Convex staking and Curve gauge balances will no longer need a manual balances refresh to be detected.
 * :feature:`-` Cowswap trades should now display the fee separately as a different event.
+* :bug:`-` If the relay reports an mev reward with same recipient and amount as block reward, do not create a duplicate block mev reward.
 
 * :release:`1.28.0 <2023-05-17>`
 * :feature:`2469` History events have now been unified under a common history events section. At the moment it features all kraken exchange events, evm events, custom imported events, block productions, staking withdrawals. Missing events retain their own sections and will be merged into the unified history in subsequent releases.